Add CA certificate example

+++ PKI/src/main/java/timejts/PKI/repository/CertificateSigningRequestRepository.java
@@ -3,10 +3,9 @@ package timejts.PKI.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import timejts.PKI.model.CertificateSigningRequest;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface CertificateSigningRequestRepository extends MongoRepository<CertificateSigningRequest, BigInteger> {
+public interface CertificateSigningRequestRepository extends MongoRepository<CertificateSigningRequest, Integer> {
 
-    Optional<CertificateSigningRequest> findById(BigInteger id);
+    Optional<CertificateSigningRequest> findById(Integer id);
 }
